[Heartbeat] Reimplement Scheduler / Fix recurring tasks with limit (#14569)
Fixes #14567

Took the opportunity to re-implement the heartbeat scheduler with an eye toward clarity as the original code was tricky to debug. While there are more LOC here, there are also additional tests and some additional features that made sense to include as well. The net result IMHO is a reduction in complexity.

This changes the behavior to start monitors using the @every X syntax immediately, rather than waiting X duration. So, we'll call this both an enhancement and a bugfix.

Structurally the code now relies more heavily on the go runtime scheduler, in combination with a new throttler package which helps constrain the maximum number of executing goroutines run by the scheduler in an efficient way. There may be another way to build the throttler using buffered go routines, but I'm on the fence whether reworking it is worth it in any way.

+++ heartbeat/config/config.go
@@ -36,7 +36,7 @@ type Config struct {
 
 // Scheduler defines the syntax of a heartbeat.yml scheduler block.
 type Scheduler struct {
-	Limit    uint   `config:"limit"  validate:"min=0"`
+	Limit    int64  `config:"limit"  validate:"min=0"`
 	Location string `config:"location"`
 }
